Remove hardcoded example components from a11y tests
